# doc/OpenAPI 3.1.yaml
openapi: 3.1.0
info:
  title: Project Chimera - S12 Gateway API
  version: 0.3.0
servers:
  - url: https://chimera.example.com/api
security:
  - bearerAuth: []
paths:
...
  /orders:
    post:
      summary: Place order (FUT/SPOT; OCO/SL supported via intent)
      operationId: ordersCreate
      security: [{ bearerAuth: [] }]
      parameters:
        - in: header
          name: X-Idempotency-Key
          required: true
          schema: { type: string, maxLength: 128 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrderCmdRequest" }
      responses:
        "200":
          description: Order result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrderResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "422": { $ref: "#/components/responses/Unprocessable" }
  /cancel:
    post:
      summary: Cancel order (optionally cascade OCO)
      operationId: ordersCancel
      security: [{ bearerAuth: [] }]
      parameters:
        - in: header
          name: X-Idempotency-Key
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CancelRequest" }
      responses:
        "200":
          description: Cancel result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CancelResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /features/recompute:
    post:
      summary: Recompute features
      operationId: featuresRecompute
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FeatureRecomputeRequest" }
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FeatureRecomputeResponse" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest: { description: Bad request }
    Unprocessable: { description: Unprocessable entity }
    NotFound: { description: Not found }
  schemas:
...
    OrderCmdRequest:
      type: object
      properties:
        intent: { $ref: "#/components/schemas/Intent" }
        tp:
          type: object
          nullable: true
          properties:
            price: { type: number }
            qty_pct: { type: number, minimum: 0, maximum: 1 }
        sl:
          type: object
          nullable: true
          properties:
            stop_price: { type: number }
            reduce_only: { type: boolean, default: true }
      required: [intent]
    Intent:
      type: object
      properties:
        intent_id: { type: string }
        market: { type: string, enum: [FUT, SPOT] }
        symbol: { type: string }
        side: { type: string, enum: [BUY, SELL] }
        qty: { type: number, exclusiveMinimum: 0 }
        type: { type: string, enum: [MARKET, LIMIT, STOP_MARKET] }
        price: { type: number }
        stop_price: { type: number }
        working_type: { type: string, enum: [MARK_PRICE, CONTRACT_PRICE], default: MARK_PRICE }
        reduce_only: { type: boolean, default: false }
        leverage: { type: integer, minimum: 1, maximum: 125 }
        isolated: { type: boolean, default: true }
        exec_policy: { type: string, enum: [MakerThenTaker, Market, OCO, LimitOnly] }
        post_only_wait_ms: { type: integer, minimum: 0, maximum: 10000, default: 3000 }
        twap:
          type: object
          properties:
            enabled: { type: boolean, default: false }
            slices: { type: integer, minimum: 1, maximum: 10, default: 3 }
            interval_ms: { type: integer, minimum: 200, maximum: 5000, default: 800 }
        oco:
          type: object
          properties:
            tp_price: { type: number }
            sl_price: { type: number }
            leg_time_in_force: { type: string, enum: [GTC, IOC], default: GTC }
        client_tags:
          type: array
          items: { type: string, maxLength: 32 }
          maxItems: 10
      required: [intent_id, market, symbol, side, qty]
    OrderResult:
      type: object
      properties:
        status: { type: string, enum: [NEW, FILLED, PARTIALLY_FILLED, ACCEPTED, CANCELED, REJECTED] }
        order_id: { type: string }
        avg_price: { type: number }
        filled_qty: { type: number }
        slippage_bps: { type: number }
        fills:
          type: array
          items:
            type: object
            properties:
              price: { type: number }
              qty: { type: number }
              fee_usdt: { type: number }
              slippage_bps: { type: number }
    CancelRequest:
      type: object
      oneOf:
        - required: [order_id]
        - required: [client_order_id]
      properties:
        order_id: { type: string }
        client_order_id: { type: string }
        cascade_oco: { type: boolean, default: true }
        reason: { type: string }
    CancelResponse:
      type: object
      properties:
        result: { type: string, enum: [CANCELLED, NOT_FOUND, FAILED] }
        order_id: { type: string }
        message: { type: string }
...
    FeatureRecomputeRequest:
      type: object
      properties:
        symbols:
          type: array
          items: { type: string }
          minItems: 1
        windows:
          type: array
          items: { type: string, enum: [1m,5m,1h,4h,1d] }
          minItems: 1
        force: { type: boolean, default: false }
      required: [symbols, windows]
    FeatureRecomputeResponse:
      type: object
      properties:
        job_id: { type: string }
        accepted: { type: boolean }
...